Add static description; re-autospec to fix dependencies
The detected spec description appears to have been the raw bytes from a
PNG file, but I cannot determine which PNG was sourced due to the
conversion to latin1 and consequent corruption of the byte stream.

Autospec no longer corrupts data like this, but to avoid adding the raw
bytes of a binary file for *this* package in the future, add a static
description.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>